+++ manta-pay/src/crypto/constraint/arkworks/mod.rs
@@ -71,232 +71,6 @@ pub mod pairing;
 #[cfg_attr(doc_cfg, doc(cfg(feature = "groth16")))]
 pub mod groth16;
 
-/// Field Element
-#[cfg_attr(
-    feature = "serde",
-    derive(Deserialize, Serialize),
-    serde(
-        bound(deserialize = "", serialize = ""),
-        crate = "manta_util::serde",
-        deny_unknown_fields,
-        try_from = "Vec<u8>"
-    )
-)]
-#[derive(Clone, Copy, Debug, Default, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct Fp<F>(
-    /// Field Element
-    #[cfg_attr(
-        feature = "serde",
-        serde(serialize_with = "serialize_field_element::<F, _>")
-    )]
-    pub F,
-)
-where
-    F: Field;
-
-impl<F> From<u128> for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn from(value: u128) -> Self {
-        Self(value.into())
-    }
-}
-
-impl<F> ToConstraintField<F> for Fp<F>
-where
-    F: PrimeField,
-{
-    #[inline]
-    fn to_field_elements(&self) -> Option<Vec<F>> {
-        self.0.to_field_elements()
-    }
-}
-
-impl<F, P> Input<P> for Fp<F>
-where
-    F: Field,
-    P: ProofSystem + ?Sized,
-    P::Input: Extend<F>,
-{
-    #[inline]
-    fn extend(&self, input: &mut P::Input) {
-        input.extend(iter::once(self.0))
-    }
-}
-
-impl<F> Decode for Fp<F>
-where
-    F: Field,
-{
-    type Error = codec::SerializationError;
-
-    #[inline]
-    fn decode<R>(reader: R) -> Result<Self, DecodeError<R::Error, Self::Error>>
-    where
-        R: Read,
-    {
-        let mut reader = codec::ArkReader::new(reader);
-        match F::deserialize(&mut reader) {
-            Ok(value) => reader
-                .finish()
-                .map(move |_| Self(value))
-                .map_err(DecodeError::Read),
-            Err(err) => Err(DecodeError::Decode(err)),
-        }
-    }
-}
-
-impl<F> Encode for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn encode<W>(&self, writer: W) -> Result<(), W::Error>
-    where
-        W: Write,
-    {
-        let mut writer = codec::ArkWriter::new(writer);
-        let _ = self.0.serialize(&mut writer);
-        writer.finish().map(move |_| ())
-    }
-}
-
-impl<F> eclair::cmp::PartialEq<Self> for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn eq(&self, rhs: &Self, _: &mut ()) -> bool {
-        PartialEq::eq(self, rhs)
-    }
-}
-
-impl<F> eclair::num::Zero for Fp<F>
-where
-    F: Field,
-{
-    type Verification = bool;
-
-    #[inline]
-    fn zero(_: &mut ()) -> Self {
-        Self(F::zero())
-    }
-
-    #[inline]
-    fn is_zero(&self, _: &mut ()) -> Self::Verification {
-        self.0.is_zero()
-    }
-}
-
-impl<F> eclair::num::One for Fp<F>
-where
-    F: Field,
-{
-    type Verification = bool;
-
-    #[inline]
-    fn one(_: &mut ()) -> Self {
-        Self(F::one())
-    }
-
-    #[inline]
-    fn is_one(&self, _: &mut ()) -> Self::Verification {
-        self.0.is_one()
-    }
-}
-
-impl<F> ConditionalSelect for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn select(bit: &Bool, true_value: &Self, false_value: &Self, _: &mut ()) -> Self {
-        if *bit {
-            *true_value
-        } else {
-            *false_value
-        }
-    }
-}
-
-impl<F> Sample for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn sample<R>(_: (), rng: &mut R) -> Self
-    where
-        R: RngCore + ?Sized,
-    {
-        Self(F::rand(rng))
-    }
-}
-
-impl<F> algebra::Group for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn add(&self, rhs: &Self, _: &mut ()) -> Self {
-        Self(self.0 + rhs.0)
-    }
-}
-
-impl<F> algebra::Ring for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn mul(&self, rhs: &Self, _: &mut ()) -> Self {
-        Self(self.0 * rhs.0)
-    }
-}
-
-impl<F> SizeLimit for Fp<F>
-where
-    F: PrimeField,
-{
-    const SIZE: usize = byte_count(<F::Params as FpParameters>::MODULUS_BITS) as usize;
-}
-
-impl<F> TryFrom<Vec<u8>> for Fp<F>
-where
-    F: Field,
-{
-    type Error = codec::SerializationError;
-
-    #[inline]
-    fn try_from(bytes: Vec<u8>) -> Result<Self, Self::Error> {
-        F::deserialize(&mut bytes.as_slice()).map(Self)
-    }
-}
-
-/// Converts `element` into its canonical byte-representation.
-#[inline]
-pub fn field_element_as_bytes<F>(element: &F) -> Vec<u8>
-where
-    F: Field,
-{
-    let mut buffer = Vec::new();
-    element
-        .serialize(&mut buffer)
-        .expect("Serialization is not allowed to fail.");
-    buffer
-}
-
-/// Uses `serializer` to serialize `element`.
-#[cfg(feature = "serde")]
-#[inline]
-fn serialize_field_element<F, S>(element: &F, serializer: S) -> Result<S::Ok, S::Error>
-where
-    F: Field,
-    S: Serializer,
-{
-    serializer.serialize_bytes(&field_element_as_bytes(element))
-}
-
 /// Synthesis Result
 pub type SynthesisResult<T = ()> = Result<T, SynthesisError>;
 
@@ -692,6 +466,7 @@ where
 /// Testing Suite
 #[cfg(test)]
 mod tests {
+    // move to constraint with R1CS
     use super::*;
     use ark_bls12_381::Fr;
     use core::iter::repeat_with;
